fix(v2.1.1): review-driven hardening of premium + Central integration
Patch release addressing 8 P0/P1 findings from the v2.1.0 self-review.

Critical correctness
  - ForwardAuditToCentralJob::failed() — was using latest()->limit(1)->update()
    which Eloquent silently treats as global UPDATE, overwriting every
    delivery row for an audit event. Fixed: first() then update on instance.
  - WebhookDelivery::markCompleted — duration_ms was permanently 0 because
    Carbon's diffInMilliseconds returns signed in 2.62+ and the args were
    reversed. Use abs() with correct arg order so it works on Carbon 2 + 3.
  - LicenseChecker::isCachedStateFresh — same Carbon signed-diff bug for
    freshness check; previously the cache was effectively pinned to the
    initial state until store-level eviction. abs() everywhere.

New-install Central-outage trap
  - Brand-new install hitting an unreachable Central was getting cached
    STATE_INVALID for the full 24h cache_ttl with no prior valid check
    to anchor grace on. New short-TTL state (5 min) with reason
    'api_unreachable_no_prior_valid' so the next request retries.

Payload bytes overflow
  - ls_webhook_deliveries.payload_bytes widened from SMALLINT (64KB) to
    INT (4.2B). Single audit events with stack-trace meta easily exceeded
    64KB. Adds migration with raw ALTER per-driver (no doctrine/dbal dep).

v1.x → v2.x live-traffic regression
  - LS_LIVE_TRAFFIC_REQUIRE_PREMIUM env (default true). False preserves
    v1.x behavior for self-hosted Reverb installs without a license.

Latency / signing edge cases
  - Cache-only license reads in hot paths: new Shield::licenseStateCached()
    + LicenseChecker::cachedState() / isPremiumCached(). Navbar badge and
    AuditLogger::maybeForwardToCentral now use cache-only — never trigger
    a 10s synchronous HTTP to Central on cache miss.
  - Cron-scheduled hourly background license refresh keeps the cache warm.
  - CentralClient::postSigned short-circuits with skipped('no_signing_secret')
    when no secret is configured (was burning 3 retries per audit event).
  - CentralClient::ping() URL derivation rewritten — parse_url-based path
    swap instead of preg_replace anchored regex (handles trailing slash,
    query strings, non-/api paths).

Quality of life
  - minutesToCron uses divisors of 60 for valid cron expressions; rounds
    up for non-divisors instead of generating "0,7,14,..." with hour-wrap
    drift.
  - broadcastWhen() logs feature-gate failures so silent broadcast loss
    is diagnosable.

+++ database/migrations/2026_05_28_000001_create_ls_webhook_deliveries_table.php
@@ -33,7 +33,10 @@ return new class extends Migration
             $table->string('operation', 32)->index();         // webhook_ingest | heartbeat | test_ping | webhook_ingest_batch
             $table->string('target_url', 255);
             $table->string('payload_hash', 64)->index();      // sha256 of the raw signed body
-            $table->unsignedSmallInteger('payload_bytes')->default(0);
+            // INT (max 4.2B) — single payloads can exceed 64KB with stack
+            // traces, and batch ingest pushes can be megabytes. The earlier
+            // SMALLINT either threw on strict MySQL or silently clamped.
+            $table->unsignedInteger('payload_bytes')->default(0);
             $table->unsignedSmallInteger('batch_size')->default(1);
 
             // Outcome
